Replace unsigned long with unsigned long long for consistency.

+++ cache.h
@@ -10,7 +10,7 @@ struct lru {
 };
 
 struct block {
-	unsigned long tag;
+	unsigned long long tag;
 	int valid;
 	int dirty;
 };
@@ -27,13 +27,13 @@ struct cache {
 	int block_bits;
 	int block_index_bits;
 	struct {
-		unsigned long requests;
-		unsigned long hits;
-		unsigned long kickouts;
-		unsigned long dirty_kickouts;
-		unsigned long flush_kickouts;
-		unsigned long reads;
-		unsigned long writes;
+		unsigned long long requests;
+		unsigned long long hits;
+		unsigned long long kickouts;
+		unsigned long long dirty_kickouts;
+		unsigned long long flush_kickouts;
+		unsigned long long reads;
+		unsigned long long writes;
 	} cache_stats;
 	struct block *buf;	/* Indexed by cache index then way */
 	struct lru **lrus;
@@ -43,17 +43,17 @@ struct cache {
 
 
 struct stat_struct {
-	unsigned long reads;
-	unsigned long writes;
-	unsigned long insts;
-	unsigned long read_cycles;
-	unsigned long write_cycles;
-	unsigned long inst_cycles;
-	unsigned long flush_time;
-	unsigned long ideal_cycles;
-	unsigned long ideal_misaligned;
-	unsigned long flushes;
-	unsigned long invalidates;
+	unsigned long long reads;
+	unsigned long long writes;
+	unsigned long long insts;
+	unsigned long long read_cycles;
+	unsigned long long write_cycles;
+	unsigned long long inst_cycles;
+	unsigned long long flush_time;
+	unsigned long long ideal_cycles;
+	unsigned long long ideal_misaligned;
+	unsigned long long flushes;
+	unsigned long long invalidates;
 };
 
 extern struct cache l1_d;
@@ -64,17 +64,17 @@ extern struct stat_struct stats;
 
 void init_cache(struct cache *cache);
 
-int dispatch_write(struct cache *cache, unsigned long addr, int bytes);
-int dispatch_read(struct cache *cache, unsigned long addr, int bytes);
-int cache_write(struct cache *cache, unsigned long addr);
-int cache_read(struct cache *cache, unsigned long addr);
+int dispatch_write(struct cache *cache, unsigned long long addr, int bytes);
+int dispatch_read(struct cache *cache, unsigned long long addr, int bytes);
+int cache_write(struct cache *cache, unsigned long long addr);
+int cache_read(struct cache *cache, unsigned long long addr);
 void l2_l1_transfer(struct cache *l1, struct cache *l2, int l2_transfer_time, int l2_bus_width);
 void print_cache(struct cache *cache);
-unsigned long cache_flush(struct cache *cache);
+unsigned long long cache_flush(struct cache *cache);
 
-static inline unsigned long int log_2(unsigned long int x)
+static inline unsigned long long int log_2(unsigned long long int x)
 {
-	unsigned long int i = 0;
+	unsigned long long int i = 0;
 	while (x != 1) { x >>= 1; i++; }
 	return i;
 }
